Backend: - Update Get Device Data API to process data in buckets based on device dashboard panel configuration   - Average aggregation for numeric values   - Last-state aggregation for text values   - Count aggregation for list values
Frontend:
- Add interactive Line/Area chart with zoom, drag, datapoint highlight, and detailed data display
- Implement filter options:
  - Live mode: display real-time data from MQTT via WebSocket
  - Range mode: predefined ranges with granularity (last min → sec, last hour → min, last day → hour, etc.)
  - Custom mode: configurable range and granularity with range limitations
- Update Meter chart to use ECharts
- Add Pie chart with custom legend (supports user input during updates)
- Update Area chart with custom legend due to ECharts limitations
- Update Data Panel to assign custom RGB colors for each data and register cases for text/list data
- Enable Line/Area chart single and multiline mode for list values
- Add Bar chart with race bar toggle
- Add Stat chart to display all value types

Database / Backend Changes:
- Update Device DB table and ingestion pipeline to support float values for data interval

+++ backend/app/models/device.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, BigInteger, DateTime, String, ForeignKey, Boolean, Integer, JSON, Enum
+from sqlalchemy import Column, BigInteger, DateTime, String, ForeignKey, Boolean, Integer, JSON, Enum, Float
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 from app.core.database import Base
@@ -21,7 +21,7 @@ class Device(Base):
     )
     mobile_number = Column(String, nullable=True)
     sim_id = Column(String, nullable=True)
-    data_interval = Column(Integer, default=60, nullable=False) # in seconds
+    data_interval = Column(Float, default=60.0, nullable=False) # in seconds
     dashboard_config = Column(JSON, nullable=True, default=dict) # JSON field for dashboard configuration
 
     is_active = Column(Boolean, default=True, nullable=False)
